[MouseGuard] Rename Deceiver to Manipulator
In line with the 2nd edition Mouse Guard character sheets. I tried to
update the translations files to add translations for "manipulator" . To
do this I used Google Translate for any of the translation files that
did not seem to just contain English anyway. The exception is
translations/zu.json, for which I was unsure of the format.